updated naming, readme, authors, environment, reqs
