+++ AUTHORS.rst
@@ -1,0 +1,7 @@
+Authors
+=======
+
+(alphabetic order)
+
+ * Amedeo Ceruti
+ * Jerry Lambert
