+++ data/Example/timeseries.csv
@@ -1,0 +1,25 @@
+;Heat Demand
+00:00;0.5284
+01:00;0.5284
+02:00;0.5313
+03:00;0.5463
+04:00;0.5672
+05:00;0.6239
+06:00;0.7463
+07:00;1
+08:00;0.9254
+09:00;0.9134
+10:00;0.8836
+11:00;0.8627
+12:00;0.8119
+13:00;0.797
+14:00;0.803
+15:00;0.8209
+16:00;0.8507
+17:00;0.8985
+18:00;0.9075
+19:00;0.9075
+20:00;0.8925
+21:00;0.8597
+22:00;0.7731
+23:00;0.6119
