+++ requirements.txt
@@ -1,0 +1,34 @@
+Bottleneck==1.3.5
+contourpy==1.2.0
+cycler==0.11.0
+fonttools==4.25.0
+kiwisolver==1.4.4
+matplotlib==3.8.0
+mkl-fft==1.3.8
+mkl-random==1.2.4
+mkl-service==2.4.0
+munkres==1.1.4
+networkx==3.1
+numexpr==2.8.7
+numpy==1.26.2
+packaging==23.1
+pandas==2.1.4
+Pillow==10.0.1
+pip==23.3.1
+ply==3.11
+pyarrow==11.0.0
+Pyomo==6.7.0
+pyparsing==3.0.9
+PyQt5==5.15.10
+PyQt5-sip==12.13.0
+python-dateutil==2.8.2
+python-louvain==0.15
+pytz==2023.3.post1
+scipy==1.11.4
+seaborn==0.12.2
+setuptools==68.0.0
+sip==6.7.12
+six==1.16.0
+tornado==6.3.3
+tzdata==2023.3
+wheel==0.41.2
