+++ results/sts/Example/topotherm_sts_vertices/optimization.log
@@ -1,0 +1,60 @@
+
+Gurobi 10.0.3 (win64) logging started Wed Dec 13 09:50:12 2023
+
+Set parameter LogFile to value "results/sts/Example/topotherm_sts_vertices/optimization.log"
+Gurobi Optimizer version 10.0.3 build v10.0.3rc0 (win64)
+
+CPU model: Intel(R) Xeon(R) W-2295 CPU @ 3.00GHz, instruction set [SSE2|AVX|AVX2|AVX512]
+Thread count: 18 physical cores, 36 logical processors, using up to 18 threads
+
+Optimize a model with 12765 rows, 11840 columns and 40423 nonzeros
+Model fingerprint: 0x1a8b4957
+Variable types: 7894 continuous, 3946 integer (3946 binary)
+Coefficient statistics:
+  Matrix range     [3e-02, 7e+04]
+  Objective range  [2e-03, 4e+04]
+  Bounds range     [1e+00, 7e+04]
+  RHS range        [4e-01, 1e+04]
+Presolve removed 10232 rows and 9237 columns
+Presolve time: 0.22s
+Presolved: 2533 rows, 2603 columns, 8209 nonzeros
+Variable types: 1265 continuous, 1338 integer (1338 binary)
+
+Root relaxation: objective 4.021422e+06, 1958 iterations, 0.07 seconds (0.04 work units)
+
+    Nodes    |    Current Node    |     Objective Bounds      |     Work
+ Expl Unexpl |  Obj  Depth IntInf | Incumbent    BestBd   Gap | It/Node Time
+
+     0     0 4021422.19    0  552          - 4021422.19      -     -    0s
+H    0     0                    4713571.6335 4021422.19  14.7%     -    0s
+     0     0 4244329.20    0  934 4713571.63 4244329.20  10.0%     -    0s
+H    0     0                    4700105.2936 4244329.20  9.70%     -    0s
+     0     0 4408379.86    0  668 4700105.29 4408379.86  6.21%     -    0s
+     0     0 4525811.45    0  510 4700105.29 4525811.45  3.71%     -    0s
+     0     0 4525829.81    0  506 4700105.29 4525829.81  3.71%     -    0s
+H    0     0                    4681906.8525 4592051.15  1.92%     -    1s
+     0     0 4592051.15    0  581 4681906.85 4592051.15  1.92%     -    1s
+     0     0 4595414.26    0  531 4681906.85 4595414.26  1.85%     -    1s
+     0     0 4595439.98    0  531 4681906.85 4595439.98  1.85%     -    1s
+H    0     0                    4648728.7028 4595439.98  1.15%     -    1s
+     0     0 4610094.84    0  706 4648728.70 4610094.84  0.83%     -    1s
+
+Cutting planes:
+  Learned: 1
+  Cover: 370
+  Implied bound: 495
+  Clique: 1125
+  MIR: 14
+  Flow cover: 252
+  GUB cover: 33
+  RLT: 1
+  Relax-and-lift: 216
+  PSD: 5
+
+Explored 1 nodes (5022 simplex iterations) in 1.45 seconds (0.67 work units)
+Thread count was 18 (of 36 available processors)
+
+Solution count 5: 4.64873e+06 4.64873e+06 4.68191e+06 ... 4.71357e+06
+
+Optimal solution found (tolerance 1.00e-02)
+Best objective 4.648728702832e+06, best bound 4.610094839890e+06, gap 0.8311%
