+++ results/sts/Example/topotherm_sts_vertices/solver.csv
@@ -1,0 +1,5 @@
+;0;unit
+Termination condition;optimal;-
+Wall time;1.4570000171661377;s
+Time;2.2921011447906494;s
+Objective;4648728.702831668;-
