+++ run_sts.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+@author: Jerry Lambert (jerry.lambert@tum.de); Amedeo Ceruti (amedeo.ceruti@tum.de)
+
+This file is used to optimize district heating systems with the tool topotherm
+of the Chair of Energy Systems.
+"""
+
+import topotherm as tt
+
+RUNID = ["Example"]
+PATH_TO_DATA = 'data/'
+
+tt.mode
